+++ frontend/src/contexts/AppContext.ts
@@ -1,53 +1,14 @@
-import {
-  createContext,
-  type Dispatch,
-  useEffect,
-  useReducer,
-  useState,
-  type SetStateAction,
-  useMemo,
-} from 'react';
+import { createContext, type Dispatch, useEffect, useState, type SetStateAction } from 'react';
 import { type Page } from '../basic/NavBar';
 import { type OpenDialogs } from '../basic/MainDialogs';
 
-import {
-  type Book,
-  defaultMaker,
-  type Maker,
-  Robot,
-  Garage,
-  Settings,
-  type Favorites,
-  Coordinator,
-  type Exchange,
-  type Order,
-  type PublicOrder,
-  type Limits,
-  defaultExchange,
-  type Federation,
-  type Version,
-  type LimitList,
-  type Origin,
-} from '../models';
+import { Settings, type Version, type Origin, Favorites } from '../models';
 
-import { apiClient } from '../services/api';
-import { systemClient } from '../services/System';
-import {
-  federationLottery,
-  getClientVersion,
-  getHost,
-  hexToBase91,
-  validateTokenEntropy,
-} from '../utils';
-import { sha256 } from 'js-sha256';
+import { getClientVersion, getHost } from '../utils';
 
 import defaultFederation from '../../static/federation.json';
-import { updateExchangeInfo } from '../models/Exchange.model';
 import { createTheme, type Theme } from '@mui/material/styles';
 import i18n from '../i18n/Web';
-import { compareUpdateLimit } from '../models/Limit.model';
-import { getEndpoint } from '../models/Coordinator.model';
-import { gridFilterActiveItemsLookupSelector } from '@mui/x-data-grid';
 
 const getWindowSize = function (fontSize: number): { width: number; height: number } {
   // returns window size in EM units
@@ -57,173 +18,15 @@ const getWindowSize = function (fontSize: number): { width: number; height: numb
   };
 };
 
-const getHostUrl = (network = 'mainnet'): { hostUrl: string; origin: Origin } => {
-  let host = '';
-  let protocol = '';
-  let origin: Origin = 'onion';
-  if (window.NativeRobosats === undefined) {
-    host = getHost();
-    protocol = location.protocol;
-  } else {
-    host = defaultFederation.exp[network].Onion;
-    protocol = 'http:';
-  }
-  const hostUrl = `${protocol}//${host}`;
-  if (window.NativeRobosats !== undefined || host.includes('.onion')) {
-    origin = 'onion';
-  } else if (host.includes('i2p')) {
-    origin = 'i2p';
-  } else {
-    origin = 'clearnet';
-  }
-
-  return { hostUrl, origin };
-};
-
-export const { hostUrl, origin } = getHostUrl();
-
-// Refresh delays (ms) according to Order status
-const statusToDelay = [
-  3000, // 'Waiting for maker bond'
-  35000, // 'Public'
-  180000, // 'Paused'
-  3000, // 'Waiting for taker bond'
-  999999, // 'Cancelled'
-  999999, // 'Expired'
-  8000, // 'Waiting for trade collateral and buyer invoice'
-  8000, // 'Waiting only for seller trade collateral'
-  8000, // 'Waiting only for buyer invoice'
-  10000, // 'Sending fiat - In chatroom'
-  10000, // 'Fiat sent - In chatroom'
-  100000, // 'In dispute'
-  999999, // 'Collaboratively cancelled'
-  10000, // 'Sending satoshis to buyer'
-  60000, // 'Sucessful trade'
-  30000, // 'Failed lightning network routing'
-  300000, // 'Wait for dispute resolution'
-  300000, // 'Maker lost dispute'
-  300000, // 'Taker lost dispute'
-];
-
 export interface SlideDirection {
   in: 'left' | 'right' | undefined;
   out: 'left' | 'right' | undefined;
 }
 
-export interface fetchRobotProps {
-  coordinator?: Coordinator;
-  newKeys?: { encPrivKey: string; pubKey: string };
-  newToken?: string;
-  slot?: number;
-  isRefresh?: boolean;
-}
-
 export type TorStatus = 'NOTINIT' | 'STARTING' | '"Done"' | 'DONE';
 
-const initialFederation: Federation = Object.entries(defaultFederation).reduce(
-  (acc, [key, value]) => {
-    acc[key] = new Coordinator(value);
-    return acc;
-  },
-  {},
-);
-
-export interface ActionFederation {
-  type:
-    | 'reset'
-    | 'enable'
-    | 'disable'
-    | 'updateBook'
-    | 'updateLimits'
-    | 'updateInfo'
-    | 'updateRobot';
-  payload: any; // TODO
-}
-
-const reduceFederation = (state: Federation, action: ActionFederation): Federation => {
-  switch (action.type) {
-    case 'reset':
-      return initialFederation;
-    case 'enable':
-      return {
-        ...state,
-        [action.payload.shortAlias]: {
-          ...state[action.payload.shortAlias],
-          enabled: true,
-        },
-      };
-    case 'disable':
-      return {
-        ...state,
-        [action.payload.shortAlias]: {
-          ...state[action.payload.shortAlias],
-          enabled: false,
-          info: undefined,
-          orders: [],
-          limits: undefined,
-        },
-      };
-    case 'updateBook':
-      return {
-        ...state,
-        [action.payload.shortAlias]: {
-          ...state[action.payload.shortAlias],
-          book: action.payload.book,
-          loadingBook: action.payload.loadingBook,
-        },
-      };
-    case 'updateLimits':
-      return {
-        ...state,
-        [action.payload.shortAlias]: {
-          ...state[action.payload.shortAlias],
-          limits: action.payload.limits,
-          loadingLimits: action.payload.loadingLimits,
-        },
-      };
-    case 'updateInfo':
-      return {
-        ...state,
-        [action.payload.shortAlias]: {
-          ...state[action.payload.shortAlias],
-          info: action.payload.info,
-          loadingInfo: action.payload.loadingInfo,
-        },
-      };
-    case 'updateRobot':
-      return {
-        ...state,
-        [action.payload.shortAlias]: {
-          ...state[action.payload.shortAlias],
-          robot: action.payload.robot,
-          loadingRobot: action.payload.loadingRobot,
-        },
-      };
-    default:
-      throw new Error(`Unhandled action type: ${String(action.type)}`);
-  }
-};
-
-const totalCoordinators = Object.keys(initialFederation).length;
-
-const initialBook: Book = {
-  orders: [],
-  loading: true,
-  loadedCoordinators: 0,
-  totalCoordinators,
-};
-
-const initialLimits: Limits = {
-  list: [],
-  loading: true,
-  loadedCoordinators: 0,
-  totalCoordinators,
-};
-
-const initialExchange: Exchange = { ...defaultExchange, totalCoordinators };
-
-const entryPage: Page | '' | 'index.html' =
-  window.NativeRobosats === undefined ? window.location.pathname.split('/')[1] : '';
+const entryPage: Page =
+  window.NativeRobosats === undefined ? (window.location.pathname.split('/')[1] as Page) : 'none';
 
 export const closeAll = {
   more: false,
@@ -252,10 +55,35 @@ const makeTheme = function (settings: Settings): Theme {
   return theme;
 };
 
-export interface CurrentOrder {
-  shortAlias: string | null;
-  id: number | null;
-}
+const getHostUrl = (network = 'mainnet'): string => {
+  let host = '';
+  let protocol = '';
+  if (window.NativeRobosats === undefined) {
+    host = getHost();
+    protocol = location.protocol;
+  } else {
+    host = defaultFederation.exp[network].Onion;
+    protocol = 'http:';
+  }
+  const hostUrl = `${protocol}//${host}`;
+
+  return hostUrl;
+};
+
+const getOrigin = (network = 'mainnet'): Origin => {
+  let host = getHostUrl(network);
+  let origin: Origin = 'onion';
+
+  if (window.NativeRobosats !== undefined || host.includes('.onion')) {
+    origin = 'onion';
+  } else if (host.includes('i2p')) {
+    origin = 'i2p';
+  } else {
+    origin = 'clearnet';
+  }
+
+  return origin;
+};
 
 export interface WindowSize {
   width: number;
@@ -263,70 +91,78 @@ export interface WindowSize {
 }
 
 export interface UseAppStoreType {
-  theme: Theme;
+  theme?: Theme;
   torStatus: TorStatus;
   settings: Settings;
   setSettings: Dispatch<SetStateAction<Settings>>;
-  book: Book;
-  setBook: Dispatch<SetStateAction<Book>>;
-  federation: Federation;
-  dispatchFederation: Dispatch<ActionFederation>;
-  sortedCoordinators: string[];
-  garage: Garage;
-  setGarage: Dispatch<SetStateAction<Garage>>;
-  currentSlot: number;
-  setCurrentSlot: Dispatch<SetStateAction<number>>;
-  fetchCoordinatorInfo: (coordinator: Coordinator) => Promise<void>;
-  fetchFederationBook: () => void;
-  limits: Limits;
-  setLimits: Dispatch<SetStateAction<Limits>>;
-  fetchFederationLimits: () => void;
-  maker: Maker;
-  setMaker: Dispatch<SetStateAction<Maker>>;
-  clearOrder: () => void;
-  robot: Robot;
-  setRobot: Dispatch<SetStateAction<Robot>>;
-  fetchFederationRobot: (props: fetchRobotProps) => void;
-  exchange: Exchange;
-  setExchange: Dispatch<SetStateAction<Exchange>>;
-  focusedCoordinator: string;
-  setFocusedCoordinator: Dispatch<SetStateAction<string>>;
-  fav: Favorites;
-  setFav: Dispatch<SetStateAction<Favorites>>;
-  order: Order | undefined;
-  setOrder: Dispatch<SetStateAction<Order | undefined>>;
-  badOrder: string | undefined;
-  setBadOrder: Dispatch<SetStateAction<string | undefined>>;
-  setDelay: Dispatch<SetStateAction<number>>;
   page: Page;
-  setAvatarLoaded: Dispatch<SetStateAction<boolean>>;
-  avatarLoaded: boolean;
   setPage: Dispatch<SetStateAction<Page>>;
   slideDirection: SlideDirection;
   setSlideDirection: Dispatch<SetStateAction<SlideDirection>>;
-  currentOrder: CurrentOrder;
-  setCurrentOrder: Dispatch<SetStateAction<CurrentOrder>>;
   navbarHeight: number;
   open: OpenDialogs;
   setOpen: Dispatch<SetStateAction<OpenDialogs>>;
-  windowSize: WindowSize;
+  windowSize?: WindowSize;
   clientVersion: {
     semver: Version;
     short: string;
     long: string;
   };
+  origin: Origin;
+  hostUrl: string;
+  fav: Favorites;
+  setFav: Dispatch<SetStateAction<Favorites>>;
 }
+
+export const initialAppContext: UseAppStoreType = {
+  theme: undefined,
+  torStatus: 'NOTINIT',
+  settings: new Settings(),
+  setSettings: () => {},
+  page: entryPage,
+  setPage: () => {},
+  slideDirection: {
+    in: undefined,
+    out: undefined,
+  },
+  setSlideDirection: () => {},
+  navbarHeight: 2.5,
+  open: closeAll,
+  setOpen: () => {},
+  windowSize: undefined,
+  origin: getOrigin(),
+  hostUrl: getHostUrl(),
+  clientVersion: getClientVersion(),
+  fav: { type: null, currency: 0, mode: 'fiat' },
+  setFav: () => {},
+};
+
+export const AppContext = createContext<UseAppStoreType>(initialAppContext);
 
 export const useAppStore = (): UseAppStoreType => {
   // State provided right at the top level of the app. A chaotic bucket of everything.
   // Contains app-wide state and functions. Triggers re-renders on the full tree often.
 
-  const [settings, setSettings] = useState<Settings>(() => {
-    return new Settings();
-  });
+  // All app data structured
+  const navbarHeight = initialAppContext.navbarHeight;
+  const clientVersion = initialAppContext.clientVersion;
+  const hostUrl = initialAppContext.hostUrl;
+  const origin = initialAppContext.origin;
+
+  const [settings, setSettings] = useState<Settings>(initialAppContext.settings);
   const [theme, setTheme] = useState<Theme>(() => {
     return makeTheme(settings);
   });
+  const [torStatus, setTorStatus] = useState<TorStatus>(initialAppContext.torStatus);
+  const [page, setPage] = useState<Page>(initialAppContext.page);
+  const [slideDirection, setSlideDirection] = useState<SlideDirection>(
+    initialAppContext.slideDirection,
+  );
+  const [open, setOpen] = useState<OpenDialogs>(initialAppContext.open);
+  const [windowSize, setWindowSize] = useState<WindowSize>(() =>
+    getWindowSize(theme.typography.fontSize),
+  );
+  const [fav, setFav] = useState<Favorites>(initialAppContext.fav);
 
   useEffect(() => {
     setTheme(makeTheme(settings));
@@ -335,59 +171,6 @@ export const useAppStore = (): UseAppStoreType => {
   useEffect(() => {
     void i18n.changeLanguage(settings.language);
   }, []);
-
-  // All app data structured
-  const [torStatus, setTorStatus] = useState<TorStatus>('NOTINIT');
-  const [book, setBook] = useState<Book>(initialBook);
-  const [limits, setLimits] = useState<Limits>(initialLimits);
-  const [garage, setGarage] = useState<Garage>(() => {
-    return new Garage();
-  });
-  const [currentSlot, setCurrentSlot] = useState<number>(() => {
-    return garage.slots.length - 1;
-  });
-  const [robot, setRobot] = useState<Robot>(() => {
-    return new Robot(garage.slots[currentSlot].robot);
-  });
-  const [maker, setMaker] = useState<Maker>(defaultMaker);
-  const [avatarLoaded, setAvatarLoaded] = useState<boolean>(false);
-  const [exchange, setExchange] = useState<Exchange>(initialExchange);
-  const [federation, dispatchFederation] = useReducer(reduceFederation, initialFederation);
-  const sortedCoordinators = useMemo(() => {
-    const sortedCoordinators = federationLottery(federation);
-    setMaker((maker) => {
-      return { ...maker, coordinator: sortedCoordinators[0] };
-    }); // default MakerForm coordinator is decided via sorted lottery
-    return sortedCoordinators;
-  }, []);
-
-  const [focusedCoordinator, setFocusedCoordinator] = useState<string>(sortedCoordinators[0]);
-  const [fav, setFav] = useState<Favorites>({ type: null, currency: 0, mode: 'fiat' });
-
-  const [delay, setDelay] = useState<number>(60000);
-  const [timer, setTimer] = useState<NodeJS.Timer | undefined>(() =>
-    setInterval(() => null, delay),
-  );
-  const [order, setOrder] = useState<Order | undefined>(undefined);
-  const [badOrder, setBadOrder] = useState<string | undefined>(undefined);
-
-  const [page, setPage] = useState<Page>(
-    entryPage === '' || entryPage === 'index.html' ? 'robot' : entryPage,
-  );
-  const [slideDirection, setSlideDirection] = useState<SlideDirection>({
-    in: undefined,
-    out: undefined,
-  });
-  const [currentOrder, setCurrentOrder] = useState<CurrentOrder>({ shortAlias: null, id: null });
-
-  const navbarHeight = 2.5;
-  const clientVersion = getClientVersion();
-
-  const [open, setOpen] = useState<OpenDialogs>(closeAll);
-
-  const [windowSize, setWindowSize] = useState<WindowSize>(() =>
-    getWindowSize(theme.typography.fontSize),
-  );
 
   useEffect(() => {
     window.addEventListener('torStatus', (event) => {
@@ -414,16 +197,6 @@ export const useAppStore = (): UseAppStoreType => {
   }, []);
 
   useEffect(() => {
-    // On bitcoin network change we reset book, limits and federation info and fetch everything again
-    setBook(initialBook);
-    setLimits(initialLimits);
-    dispatchFederation({ type: 'reset' });
-    fetchFederationBook();
-    fetchFederationInfo();
-    fetchFederationLimits();
-  }, [settings.network]);
-
-  useEffect(() => {
     setWindowSize(getWindowSize(theme.typography.fontSize));
   }, [theme.typography.fontSize]);
 
@@ -431,456 +204,23 @@ export const useAppStore = (): UseAppStoreType => {
     setWindowSize(getWindowSize(theme.typography.fontSize));
   };
 
-  // fetch Limits
-  const fetchCoordinatorLimits = async (coordinator: Coordinator): Promise<void> => {
-    const { url, basePath } = getEndpoint({
-      network: settings.network,
-      coordinator,
-      origin,
-      selfHosted: settings.selfhostedClient,
-      hostUrl,
-    });
-
-    const limits = await apiClient
-      .get(url, `${basePath}/api/limits/`)
-      .then((data) => {
-        return data;
-      })
-      .catch(() => {
-        return undefined;
-      });
-    dispatchFederation({
-      type: 'updateLimits',
-      payload: { shortAlias: coordinator.shortAlias, limits, loadingLimits: false },
-    });
-  };
-
-  const fetchFederationLimits = function (): void {
-    Object.entries(federation).map(([shortAlias, coordinator]) => {
-      if (coordinator.enabled === true) {
-        // set limitLoading=true
-        dispatchFederation({
-          type: 'updateLimits',
-          payload: { shortAlias, limits: coordinator.limits, loadingLimits: true },
-        });
-        // fetch new limits
-        void fetchCoordinatorLimits(coordinator);
-      }
-      return null; // Object.entries() expect a return
-    });
-  };
-
-  // fetch Books
-  const fetchCoordinatorBook = async (coordinator: Coordinator): Promise<void> => {
-    const { url, basePath } = getEndpoint({
-      network: settings.network,
-      coordinator,
-      origin,
-      selfHosted: settings.selfhostedClient,
-      hostUrl,
-    });
-
-    const book = await apiClient
-      .get(url, `${basePath}/api/book/`)
-      .then((data: PublicOrder[]) => {
-        return data.not_found !== undefined ? [] : data;
-      })
-      .catch(() => {
-        return [];
-      });
-    dispatchFederation({
-      type: 'updateBook',
-      payload: { shortAlias: coordinator.shortAlias, book, loadingBook: false },
-    });
-  };
-
-  const fetchFederationBook = function (): void {
-    Object.entries(federation).map(([shortAlias, coordinator]) => {
-      if (coordinator.enabled === true) {
-        dispatchFederation({
-          type: 'updateBook',
-          payload: { shortAlias, book: coordinator.book, loadingBook: true },
-        });
-        void fetchCoordinatorBook(coordinator);
-      }
-      return null; // Object.entries() expect a return
-    });
-  };
-
-  // fetch Info
-  const fetchCoordinatorInfo = async (coordinator: Coordinator): Promise<void> => {
-    // Set loading true
-    dispatchFederation({
-      type: 'updateInfo',
-      payload: { shortAlias: coordinator.shortAlias, info: coordinator.info, loadingInfo: true },
-    });
-    // fetch and dispatch
-    const { url, basePath } = getEndpoint({
-      network: settings.network,
-      coordinator,
-      origin,
-      selfHosted: settings.selfhostedClient,
-      hostUrl,
-    });
-
-    const info = await apiClient
-      .get(url, `${basePath}/api/info/`)
-      .then((data) => {
-        return data;
-      })
-      .catch(() => {
-        return undefined;
-      });
-    dispatchFederation({
-      type: 'updateInfo',
-      payload: { shortAlias: coordinator.shortAlias, info, loadingInfo: false },
-    });
-  };
-
-  const fetchFederationInfo = function (): void {
-    Object.entries(federation).map(([shortAlias, coordinator]) => {
-      if (coordinator.enabled === true) {
-        dispatchFederation({
-          type: 'updateInfo',
-          payload: { shortAlias, info: coordinator.info, loadingInfo: true },
-        });
-        void fetchCoordinatorInfo(coordinator);
-      }
-      return null; // Object.entries() expect a return
-    });
-  };
-
-  const updateBook = (): void => {
-    setBook((book) => {
-      return { ...book, loading: true, loadedCoordinators: 0 };
-    });
-    let orders: PublicOrder[] = book.orders;
-    let loadedCoordinators: number = 0;
-    let totalCoordinators: number = 0;
-
-    sortedCoordinators.map((shortAlias: string) => {
-      if (federation[shortAlias]?.enabled === true) {
-        totalCoordinators = totalCoordinators + 1;
-        if (!federation[shortAlias].loadingBook) {
-          const existingOrders = orders.filter(
-            (order) => order.coordinatorShortAlias !== shortAlias,
-          );
-          const newOrders: PublicOrder[] = federation[shortAlias].book.map((order) => ({
-            ...order,
-            coordinatorShortAlias: shortAlias,
-          }));
-          orders = [...existingOrders, ...newOrders];
-          // orders.push.apply(existingOrders, newOrders);
-          loadedCoordinators = loadedCoordinators + 1;
-        }
-      }
-      const loading = loadedCoordinators !== totalCoordinators;
-      setBook({ orders, loading, loadedCoordinators, totalCoordinators });
-      return null; // Object.values() expects a return
-    });
-  };
-
-  const updateLimits = (): void => {
-    const newLimits: LimitList | never[] = [];
-    setLimits((limits) => {
-      return { ...limits, loadedCoordinators: 0 };
-    });
-    Object.entries(federation).map(([shortAlias, coordinator]) => {
-      if (coordinator.limits !== undefined) {
-        for (const currency in coordinator.limits) {
-          newLimits[currency] = compareUpdateLimit(
-            newLimits[currency],
-            coordinator.limits[currency],
-          );
-        }
-        setLimits((limits) => {
-          return {
-            ...limits,
-            list: newLimits,
-            loading: true,
-            loadedCoordinators: limits.loadedCoordinators + 1,
-          };
-        });
-      }
-      return null; // Object.entries expects a return
-    });
-    setLimits((limits) => {
-      return { ...limits, loading: false };
-    });
-  };
-
-  const updateExchange = (): void => {
-    const onlineCoordinators = Object.keys(federation).reduce((count, shortAlias): number => {
-      if (!federation[shortAlias]?.loadingInfo && federation[shortAlias]?.info !== undefined) {
-        return count + 1;
-      } else {
-        return count;
-      }
-    }, 0);
-    const totalCoordinators = Object.keys(federation).reduce((count, shortAlias) => {
-      return federation[shortAlias]?.enabled === true ? count + 1 : count;
-    }, 0);
-    setExchange({ info: updateExchangeInfo(federation), onlineCoordinators, totalCoordinators });
-  };
-
-  useEffect(() => {
-    updateBook();
-    updateLimits();
-    updateExchange();
-  }, [federation]);
-
-  useEffect(() => {
-    if (open.exchange) {
-      fetchFederationInfo();
-    }
-  }, [open.exchange, torStatus]);
-
-  useEffect(() => {
-    fetchFederationInfo();
-  }, []);
-
-  // Fetch current order at load and in a loop
-  useEffect(() => {
-    if (currentOrder.id != null && (page === 'order' || (order === badOrder) === undefined)) {
-      fetchOrder();
-    }
-  }, [currentOrder, page]);
-
-  useEffect(() => {
-    clearInterval(timer);
-    setTimer(setInterval(fetchOrder, delay));
-    return () => {
-      clearInterval(timer);
-    };
-  }, [delay, currentOrder, page, badOrder]);
-
-  const orderReceived = function (data: any): void {
-    if (data.bad_request !== undefined) {
-      setBadOrder(data.bad_request);
-      setDelay(99999999);
-      setOrder(undefined);
-    } else {
-      setDelay(
-        data.status >= 0 && data.status <= 18
-          ? page === 'order'
-            ? statusToDelay[data.status]
-            : statusToDelay[data.status] * 5
-          : 99999999,
-      );
-      setOrder(data);
-      setBadOrder(undefined);
-    }
-  };
-
-  const fetchOrder = function (): void {
-    if (currentOrder.shortAlias != null && currentOrder.id != null) {
-      const { url, basePath } = getEndpoint({
-        network: settings.network,
-        coordinator: federation[currentOrder.shortAlias],
-        origin,
-        selfHosted: settings.selfhostedClient,
-        hostUrl,
-      });
-      const auth = {
-        tokenSHA256: robot.tokenSHA256,
-        keys: {
-          pubKey: robot.pubKey.split('\n').join('\\'),
-          encPrivKey: robot.encPrivKey.split('\n').join('\\'),
-        },
-      };
-
-      void apiClient
-        .get(url, `${basePath}/api/order/?order_id=${currentOrder.id}`, auth)
-        .then(orderReceived)
-        .catch(orderReceived);
-    }
-  };
-
-  const clearOrder = function (): void {
-    setOrder(undefined);
-    setBadOrder(undefined);
-  };
-
-  const fetchCoordinatorRobot = function ({
-    coordinator,
-    newToken,
-    newKeys,
-    slot,
-    isRefresh = false,
-  }: fetchRobotProps): void {
-    const { url, basePath } = getEndpoint({
-      network: settings.network,
-      coordinator,
-      origin,
-      selfHosted: settings.selfhostedClient,
-      hostUrl,
-    });
-
-    const token = newToken ?? robot.token ?? '';
-
-    const { hasEnoughEntropy, bitsEntropy, shannonEntropy } = validateTokenEntropy(token);
-
-    if (!hasEnoughEntropy) {
-      return;
-    }
-
-    const tokenSHA256 = hexToBase91(sha256(token));
-    const targetSlot = slot ?? currentSlot;
-    const encPrivKey = newKeys?.encPrivKey ?? robot.encPrivKey ?? '';
-    const pubKey = newKeys?.pubKey ?? robot.pubKey ?? '';
-
-    // On first authenticated requests, pubkey and privkey are needed in header cookies
-    const auth = {
-      tokenSHA256,
-      keys: {
-        pubKey: pubKey.split('\n').join('\\'),
-        encPrivKey: encPrivKey.split('\n').join('\\'),
-      },
-    };
-
-    if (!isRefresh && coordinator) {
-      const newRobot = {
-        ...coordinator.robot,
-        loading: true,
-      };
-
-      dispatchFederation({
-        type: 'updateRobot',
-        payload: { shortAlias: coordinator.shortAlias, robot: newRobot, loadingRobot: false },
-      });
-      setRobot(newRobot);
-    }
-
-    apiClient
-      .get(url, `${basePath}/api/robot/`, auth)
-      .then((data: any) => {
-        const newRobot = {
-          nickname: data.nickname,
-          token,
-          tokenSHA256,
-          loading: false,
-          activeOrderId: data.active_order_id ?? null,
-          lastOrderId: data.last_order_id ?? null,
-          earnedRewards: data.earned_rewards ?? 0,
-          stealthInvoices: data.wants_stealth,
-          tgEnabled: data.tg_enabled,
-          tgBotName: data.tg_bot_name,
-          tgToken: data.tg_token,
-          found: data?.found,
-          last_login: data.last_login,
-          bitsEntropy,
-          shannonEntropy,
-          pubKey: data.public_key,
-          encPrivKey: data.encrypted_private_key,
-          copiedToken: Boolean(data.found),
-        };
-        if (currentOrder.id == null) {
-          setCurrentOrder({
-            id:
-              data.active_order_id !== undefined
-                ? data.active_order_id
-                : data.last_order_id !== undefined
-                ? data.last_order_id
-                : null,
-            shortAlias: coordinator?.shortAlias,
-          });
-        }
-
-        setRobot(newRobot);
-        garage.updateRobot(newRobot, targetSlot);
-        setCurrentSlot(targetSlot);
-
-        dispatchFederation({
-          type: 'updateRobot',
-          payload: { shortAlias: coordinator.shortAlias, robot: newRobot, loadingRobot: false },
-        });
-      })
-      .finally(() => {
-        systemClient.deleteCookie('public_key');
-        systemClient.deleteCookie('encrypted_private_key');
-      });
-  };
-
-  const fetchFederationRobot = function (props: fetchRobotProps): void {
-    if (!props?.isRefresh) setAvatarLoaded(false);
-
-    Object.entries(federation).map(([shortAlias, coordinator]) => {
-      if (coordinator.enabled === true) {
-        dispatchFederation({
-          type: 'updateRobot',
-          payload: { shortAlias, robot: coordinator.robot, loadingRobot: true },
-        });
-        fetchCoordinatorRobot({ ...props, coordinator });
-      }
-      return null; // Object.entries expects a return
-    });
-  };
-
-  useEffect(() => {
-    if (page !== 'robot') {
-      if (open.profile && avatarLoaded) {
-        fetchFederationRobot({ isRefresh: true }); // refresh/update existing robot
-      } else if (
-        !avatarLoaded &&
-        robot.token !== undefined &&
-        robot.encPrivKey !== undefined &&
-        robot.pubKey !== undefined
-      ) {
-        fetchFederationRobot({}); // create new robot with existing token and keys (on network and coordinator change)
-      }
-    }
-  }, [open.profile, hostUrl]);
-
   return {
     theme,
     torStatus,
     settings,
     setSettings,
-    book,
-    setBook,
-    federation,
-    dispatchFederation,
-    sortedCoordinators,
-    garage,
-    setGarage,
-    currentSlot,
-    setCurrentSlot,
-    fetchCoordinatorInfo,
-    fetchFederationBook,
-    limits,
-    setLimits,
-    fetchFederationLimits,
-    maker,
-    setMaker,
-    clearOrder,
-    robot,
-    setRobot,
-    fetchFederationRobot,
-    exchange,
-    setExchange,
-    focusedCoordinator,
-    setFocusedCoordinator,
-    fav,
-    setFav,
-    order,
-    setOrder,
-    badOrder,
-    setBadOrder,
-    setDelay,
     page,
     setPage,
     slideDirection,
     setSlideDirection,
-    currentOrder,
-    setCurrentOrder,
     navbarHeight,
     open,
     setOpen,
     windowSize,
     clientVersion,
-    setAvatarLoaded,
-    avatarLoaded,
+    hostUrl,
+    origin,
+    fav,
+    setFav,
   };
 };
-
-export const AppContext = createContext<UseAppStoreType | undefined>(undefined);
