+++ frontend/src/pro/Widgets/Maker.tsx
@@ -3,6 +3,8 @@ import { AppContext, type UseAppStoreType } from '../../contexts/AppContext';
 
 import MakerForm from '../../components/MakerForm';
 import { Paper } from '@mui/material';
+import { FederationContext, UseFederationStoreType } from '../../contexts/FederationContext';
+import { GarageContext, UseGarageStoreType } from '../../contexts/GarageContext';
 
 interface MakerWidgetProps {
   style?: React.StyleHTMLAttributes<HTMLElement>;
@@ -16,7 +18,9 @@ const MakerWidget = React.forwardRef(function Component(
   { style, className, onMouseDown, onMouseUp, onTouchEnd }: MakerWidgetProps,
   ref,
 ) {
-  const { maker, fav, limits } = useContext<UseAppStoreType>(AppContext);
+  const { fav } = useContext<UseAppStoreType>(AppContext);
+  const { limits } = useContext<UseFederationStoreType>(FederationContext);
+  const { maker } = useContext<UseGarageStoreType>(GarageContext);
   return React.useMemo(() => {
     return (
       <Paper elevation={3} style={{ padding: 8, overflow: 'auto', width: '100%', height: '100%' }}>
